+++ packages/console-types/src/constants/pagination.ts
@@ -1,0 +1,2 @@
+export const TOTAL_ITEMS_HEADER_KEY = 'x-total-items'
+export const TOTAL_PAGES_HEADER_KEY = 'x-total-pages'